requirements.txt
contourpy==1.2.1
cycler==0.12.1
fonttools==4.50.0
JSBSim==1.2.0
kiwisolver==1.4.5
matplotlib==3.8.4
numpy==1.26.4
packaging==24.0
pillow==10.3.0
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil==2.9.0.post0
pyzmq==25.1.2
six==1.16.0
